python build fix after rebase
